Tests for NYSIIS.
This originally used a corpus from the "aspell" dictionary, which
was run through a small program to generate the corresponding
NYSIIS output using the apache commons implementation. I felt
this was overkill for this particular algorithm (unlike that of
double metaphone which is way more complex), so I replaced it with
these simpler cases.